Updated parser

+++ parser_file.py
@@ -33,7 +33,7 @@ def parse_arguments():
 
     parser.add_argument("--threshold", type = int, default = 4, help = "Choose a threshold for anomaly detection")
     parser.add_argument("--percentile", type = float, default = 0.95, help = "Choose a percentile for anomaly detection")
-    parser.add_argument("--weights_overall", type = int, default = 0.5, help = "Choose a weight for anomaly detection")
+    parser.add_argument("--weights_overall", type = float, default = 0.5, help = "Choose a weight for anomaly detection")
 
     parser.add_argument("--do_multivariate", action = "store_true", help = "Whether to perform multivariate trials or not")
